# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List all posts
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, posted, failed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostListResponse"
...
  /posts/generate:
    post:
      operationId: generatePost
      tags: [posts]
      summary: Manually trigger content generation
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [image, reels, carousel]
                  default: image
                imageSource:
                  type: string
                  enum: [ai, search]
      responses:
        "200":
          description: Generated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a single post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                caption:
                  type: string
                hashtags:
                  type: string
                status:
                  type: string
                  enum: [pending, approved, rejected, posted]
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}/publish:
    post:
      operationId: publishPost
      tags: [posts]
      summary: Publish a post to Instagram
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Published post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}/approve:
    post:
      operationId: approvePost
      tags: [posts]
      summary: Approve a pending post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}/reject:
    post:
      operationId: rejectPost
      tags: [posts]
      summary: Reject a pending post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
...
components:
  schemas:
...
    Post:
      type: object
      properties:
        id:
          type: integer
        caption:
          type: string
        hashtags:
          type: string
        imageUrl:
          type: string
        videoUrl:
          type: string
          nullable: true
        mediaType:
          type: string
          enum: [image, reels, carousel]
          default: image
        mediaUrls:
          type: array
          items:
            type: string
          nullable: true
        imagePrompt:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, posted, failed]
        publishError:
          type: string
          nullable: true
        scheduledFor:
          type: string
          format: date-time
        postedAt:
          type: string
          format: date-time
          nullable: true
        instagramPostId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - caption
        - hashtags
        - imageUrl
        - mediaType
        - imagePrompt
        - status
        - scheduledFor
        - createdAt
        - updatedAt
    PostListResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
      required:
        - posts
        - total
...